requirements.txt
faiss-cpu
pymongo
aiohttp
uvicorn
numpy
pydantic
fastapi
sentence-transformers
torch
imageio
Pillow
diffusers
huggingface_hub
safetensors
